Implementation of modular design with Flask blueprints.
