+++ LabManager/templates/layout.html
@@ -5,7 +5,7 @@
         <meta name="viewport" content="width=device-width, initial-scale=1.0">
         <link rel="stylesheet" href="{{ url_for('static', filename='css/main.css') }}">
         <link href="https://fonts.googleapis.com/css?family=Oxygen|Allerta+Stencil" rel="stylesheet">
-        <link href="https://cdn.jsdelivr.net/npm/tailwindcss/dist/tailwind.min.css" rel="stylesheet">
+        <link href="https://unpkg.com/tailwindcss@^1.0.4/dist/tailwind.min.css" rel="stylesheet">
         {% block head %}{% endblock %}
         {% if title %}
             <title>{{ title }}</title>
@@ -136,11 +136,11 @@
                 {% if messages %}
                     {% for category, message in messages %}
                         {% if category == "error" %}
-                            <div class="fixed w-full bg-red-light rounded-lg text-gray-300 text-sm font-bold px-4 py-3 m-1">
+                            <div class="fixed w-full bg-red-400 rounded-lg text-gray-300 text-sm font-bold px-4 py-3 m-1 z-20">
                                 {{ message }}
                             </div>
                         {% else %}
-                            <div class="fixed w-full bg-blue-500 rounded-lg text-gray-300 text-sm font-bold px-4 py-3 m-1">
+                            <div class="fixed w-full bg-blue-500 rounded-lg text-gray-300 text-sm font-bold px-4 py-3 m-1 z-20">
                                 {{ message }}
                             </div>
                         {% endif %}
